+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.8
+---
+
+* Not yet released.
+
 1.7
 ---